+++ app/components/movie/movie.component.scss
