+++ app/index.js
@@ -5,9 +5,9 @@ import { Router, Route, IndexRoute, browserHistory } from 'react-router';
 import { syncHistoryWithStore } from 'react-router-redux';
 
 import configureStore from './store/configureStore';
-import App from './containers/App';
-import MoviesContainer from './containers/MoviesContainer';
-import LoginContainer from './containers/LoginContainer';
+import App from './containers/app/app.container';
+import MoviesContainer from './containers/movies/movies.container';
+import LoginContainer from './containers/login/login.container';
 
 const store = configureStore(undefined, browserHistory);
