+++ app/reducers/root.reducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import { routerReducer as routing } from 'react-router-redux';
-import loginReducer from './loginReducer';
-import movies from './movies';
+import loginReducer from './login.reducer';
+import movies from './movies.reducer';
 
 export default combineReducers({
   routing,
